[#260] Improve the output for failed Nix commands
Problem: The output when a command failed for deployment would show the
error message in the console output and _then_ the error message, which
was confusing.

Solution: Show the error message together with the error in an order
that makes sense. This is done by capturing the output in the error
message and printinting it in an exit error. We also use `Stdio::null`
rather than `Stdio::piped` to avoid printing to the console.